refactor: move canvas straight to views
