+++ frontend/src/features/power_unit/three/LineChart.tsx
@@ -1,9 +1,21 @@
 import useAxes from "./hooks/useAxes";
-import { ChartProps } from "./Chart2D";
 import Trace from "./Trace";
 import LinesVertical from "./LinesVertical";
 import LinesHorizontal from "./LinesHorizontal";
 import useChartSize from "./hooks/useChartSize";
+
+export interface Axis {
+  name: string;
+  type?: string;
+  min?: number;
+  max?: number;
+}
+
+export type ChartProps = {
+  trace: number[][];
+  xAxis: Axis;
+  yAxis: Axis;
+};
 
 const LineChart = ({ trace, xAxis, yAxis }: ChartProps) => {
   const { xTicks, yTicks, scaleX, scaleY } = useAxes(trace, xAxis, yAxis);
@@ -11,9 +23,9 @@ const LineChart = ({ trace, xAxis, yAxis }: ChartProps) => {
 
   return (
     <mesh position={[-0.4 * width, -0.5 * height, 0]}>
-      <Trace trace={trace} scale={[scaleX, scaleY, 1]} />
       <LinesVertical ticks={xTicks} scale={scaleX} axis={xAxis} />
       <LinesHorizontal ticks={yTicks} scale={scaleY} axis={yAxis} />
+      <Trace trace={trace} scale={[scaleX, scaleY, 1]} />
     </mesh>
   );
 };
